refactor(ReactionButton): 좋아요/싫어요 버튼 컴포넌트 분리
